requirements.txt
numpy==1.23.5
pandas==1.5.3
sentence-transformers==2.2.2
fastapi==0.68.1
uvicorn==0.15.0
torch==2.0.1
transformers==4.31.0
pydantic==1.10.12
beautifulsoup4==4.12.2
python-multipart==0.0.6
python-dotenv==1.0.0
bs4==0.0.1